Create README.md
Contract Overview
The SimpleStorage contract allows users to store their favorite number and manage a list of people along with their favorite numbers. Below are the key components and functionalities of the contract:

State Variables
uint256 myfavoriteNumber;: A state variable to store a single favorite number. If no value is provided, the default value is 0.
Person[] public listofPeople;: An array to store multiple Person structs.
mapping (string => uint256) public nameToFavoriteNumber;: A mapping to connect a person's name to their favorite number.
Struct
struct Person: A struct that holds a person's favorite number and name.
Functions
function store(uint256 _favoriteNumber) public: Stores a provided favorite number in the myfavoriteNumber state variable.
function retrieve() public view returns (uint256): Retrieves the stored favorite number. It is a view function, meaning it does not modify the state.
function addPerson(string memory _name, uint256 _favoriteNumber) public: Adds a new person to the listofPeople array and updates the nameToFavoriteNumber mapping.
Usage
Storing a Favorite Number:
Call the store function with a uint256 argument to store a favorite number.
Retrieving the Stored Number:
Call the retrieve function to get the currently stored favorite number.
Adding a Person:
Call the addPerson function with a string (name) and uint256 (favorite number) to add a person to the list and update the mapping.